moved carbon theme items around w/ burnettk
