+++ spiffworkflow-frontend/.eslintignore
@@ -1,1 +1,1 @@
-/src/themes/carbon
+/src/rjsf/carbon_theme
