+++ spiffworkflow-frontend/src/rjsf/carbon_theme/ArrayFieldItemTemplate/ArrayFieldItemTemplate.tsx
