+++ spiffworkflow-frontend/src/rjsf/carbon_theme/SubmitButton/SubmitButton.tsx
@@ -1,5 +1,4 @@
 import React from 'react';
-// import Box from '@mui/material/Box';
 // @ts-ignore
 import { Button } from '@carbon/react';
 import { SubmitButtonProps, getSubmitButtonOptions } from '@rjsf/utils';
