+++ spiffworkflow-frontend/src/routes/TaskShow.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useNavigate, useParams } from 'react-router-dom';
 import validator from '@rjsf/validator-ajv8';
 
@@ -8,85 +8,18 @@ import {
   Tabs,
   Grid,
   Column,
-  ComboBox,
   Button,
   ButtonSet,
 } from '@carbon/react';
 
-// eslint-disable-next-line import/no-named-as-default
-import Form from '../themes/carbon';
+import { Form } from '../rjsf/carbon_theme';
 import HttpService from '../services/HttpService';
 import useAPIError from '../hooks/UseApiError';
 import { modifyProcessIdentifierForPathParam } from '../helpers';
 import { EventDefinition, Task } from '../interfaces';
 import ProcessBreadcrumb from '../components/ProcessBreadcrumb';
 import InstructionsForEndUser from '../components/InstructionsForEndUser';
-
-// TODO: move this somewhere else
-function TypeaheadWidget({
-  id,
-  onChange,
-  options: { category, itemFormat },
-}: {
-  id: string;
-  onChange: any;
-  options: any;
-}) {
-  const pathForCategory = (inputText: string) => {
-    return `/connector-proxy/typeahead/${category}?prefix=${inputText}&limit=100`;
-  };
-
-  const lastSearchTerm = useRef('');
-  const [items, setItems] = useState<any[]>([]);
-  const [selectedItem, setSelectedItem] = useState<any>(null);
-  const itemFormatRegex = /[^{}]+(?=})/g;
-  const itemFormatSubstitutions = itemFormat.match(itemFormatRegex);
-
-  const itemToString = (item: any) => {
-    if (!item) {
-      return null;
-    }
-
-    let str = itemFormat;
-    itemFormatSubstitutions.forEach((key: string) => {
-      str = str.replace(`{${key}}`, item[key]);
-    });
-    return str;
-  };
-
-  const handleTypeAheadResult = (result: any, inputText: string) => {
-    if (lastSearchTerm.current === inputText) {
-      setItems(result);
-    }
-  };
-
-  const typeaheadSearch = (inputText: string) => {
-    if (inputText) {
-      lastSearchTerm.current = inputText;
-      // TODO: check cache of prefixes -> results
-      HttpService.makeCallToBackend({
-        path: pathForCategory(inputText),
-        successCallback: (result: any) =>
-          handleTypeAheadResult(result, inputText),
-      });
-    }
-  };
-
-  return (
-    <ComboBox
-      onInputChange={typeaheadSearch}
-      onChange={(event: any) => {
-        setSelectedItem(event.selectedItem);
-        onChange(itemToString(event.selectedItem));
-      }}
-      id={id}
-      items={items}
-      itemToString={itemToString}
-      placeholder={`Start typing to search for ${category}...`}
-      selectedItem={selectedItem}
-    />
-  );
-}
+import TypeaheadWidget from '../rjsf/custom_widgets/TypeaheadWidget/TypeaheadWidget';
 
 export default function TaskShow() {
   const [task, setTask] = useState<Task | null>(null);
